new jdk

+++ run_autotest/src/main/java/com/xn/test/util/RedisUtil.java
@@ -8,8 +8,6 @@ import cn.xn.user.enums.SourceType;
 import cn.xn.user.enums.SystemType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.support.ClassPathXmlApplicationContext;
 import org.springframework.stereotype.Service;
 import redis.clients.jedis.JedisCluster;
 
@@ -263,8 +261,8 @@ public class RedisUtil {
         return isExit;
     }
 
-    public int getTime(String key) {
-        int time = Math.toIntExact(jedisCluster.ttl(key));
+    public Long getTime(String key) {
+        Long time = jedisCluster.ttl(key);
         logger.warn("key:{} expire time is {}", key, time);
         return time;
     }
